ml-app/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
transformers==4.35.2
torch==2.1.0
torchvision==0.16.0
pillow==10.1.0
python-multipart==0.0.6
pydantic-settings==2.0.3
python-decouple==3.8
requests==2.31.0
loguru==0.7.3
accelerate==0.24.1